treewide: use fd_clock-based timestamping
Switch all tiles from fd_log_wallclock() and fd_tickcount() to fd_clock.

+++ src/disco/topo/fd_topo.h
@@ -200,8 +200,8 @@ struct fd_topo_tile {
       ulong  max_concurrent_connections;
       ulong  max_concurrent_handshakes;
       ushort quic_transaction_listen_port;
-      ulong  idle_timeout_millis;
-      uint   ack_delay_millis;
+      long   idle_timeout_millis;
+      long   ack_delay_millis;
       int    retry;
       char   key_log_path[ PATH_MAX ];
     } quic;
